null (missing) "matchesAssembly" in mongo means "true"
We can not modify ImportedSubmittedVariantEntity nor SubmittedVariantEntity
to reflect this default value because then the "true" value would get written
into mongo. The entities are a java representation of the DB, so the values
there can be empty, it's the classes that interpret the entities the ones that
should have this knowledge.