add test for a corner case (repeated merges)
for the rs136611820 test we are going to need to be more specific on
when not to issue a merge operation, if the accession is already merged.

This commit changes the condition to issue the merge operation. It was
"if the accession is not merged", now it is
"if the combination of (accession, hash and mergeInto) is not merged"