make assembly name optional
if there's only one assembly name in the contig_info table, use
that one. Otherwise raise an exception, and the parameter will
have to be provided manually (-n,--assembly_name).

If there are several ones you will be able to see which ones and
copy one of them from the program output.