Decided not to sort
quick benchmarks show that the sorting might not use the
accession indexes properly, as 90% of the step time is spent
before any result is read.

for apple (140K deprecated RS IDs) the sorted cursor step lasted
5 minutes, while the unsorted cursor step was 10 seconds, and a
bash sort was around 5 seconds.

The paging strategy with sorting was 4.5 hours, but the unsorted
paging strategy was not tried.