change configuration to use ContigMongoReader
There's no need for ContigStepPreviousBuildConfigurationTest
anymore, as the contig reader (and any other reader) doesn't care
about which build the variants are. Even in the exceptional case
where there are several builds for the same assembly, we loaded
them uniformingly into the same assembly in mongo.

Also, with the removal of the usage of the postgres contig reader
(SubSnpNoHgvsContigReader) the release module doesn't need any
connection to postgres, so the datasource part can be removed. We
are keeping the jobRepository datasource, though, to resume failed
jobs.