revisit eva-accession-import contig replacement
The previous checks were overly strict. it wouldn't perform a
"chr to genbank" replacement if the refseq was "na" and
non-identical. That case can safely be replaced.

chr to genbank replacement is suspicious only if genbank and refseq
are non-identical and refseq is present. This is still done if
forceImporty=true is provided.