show bug: getOrCreate(emptylist) should not fail
This can happen if asked to accession variants that are present
in dbsnp. The service only accessions the ones that are not in
dbsnp, removing the present ones from the initial list, resulting
in an empty list to be accessioned in the EVA side, resulting in
an insert of an empty list, which mongo forbids.