use system sort for the list of contigs
Of course this is platform dependent, but given that we have some
species with lots of contigs (e.g. almost 0.5M for rat or turkey),
we can either:

- keep the hacky implementation of this commit.
- evaluate memory performance of: loading the file into
a list, sort the list in memory, write the list. For at least a
couple million of 20 bytes strings.

If we go for the second option, the changes have to be done in the
same place as the current system sort.